WIP: introduce Visualizer for rendering in notebooks.
This change is a Work In Progress.

It should be possible to render Mesa's javascript visualizations
directly in jupyter notebook cells, where the model has been defined in
the notebook itself.

This introduces a helper Visualizer class which will inject the required
javascript into the notebook, deal with running the model and, and
sending the updated state data to the notebook.

This change is dependent upon #239

fixes #246