Adding gradient support for agents in GridVisualization
I've renamed the Color parameter to Colors and made it an array. If the list consists of only one color then the shape is filled only with that color, otherwise it fills it with a gradient created with that colors. 
I've also added an optional parameter called stroke_color. If it is not provided than it becomes the first given color in the colors array.