Added new classes with support for evolutionary models
EvoAgent extends Agent to deal with energy, age, death, reproduction and
genome

EvoModel extends Model to deal with death and reproduction in the
population. It also supports mixed populations with agents of different types.
EvoModel requires a subclass of SelectionModel.

SelectionModel determines the modalities of death and reproduction.
The base class does nothing.
Subclass ProbabilisticSelection agents die with a given probability and
reproduce with another, determined by model parameters. If the agent specifies
individual probabilities, they are used instead of the global probabilities.
Subclass RouletteSelection implements the roulette-wheel selection with
stochastic acceptance: a fraction of the population is considered for death
and reproduction, each agent obtains a weight based on its characteristics
(age/energy) and a model parameter (max_age/max_energy) and it dies/reproduces
with a probability proportional to the weight.

SelectionModel requires a subclass of ReproductioModel
ReproductionModel deals with the reproduction. The base class clones the agent.
Subclass AsexualReproduction create an offspring that is derived from the
agent, with a mutated dna.
Subclass SexualReproduction finds a suitable partner in the population and
reproduces by crossover and mutation.