- Moved model initialization into the iterations loop to ensure that models are also properly initialized when starting a new iteration.
- The deepcopy ensures that on model initialization the kwargs are always
the original ones specified by the user.
The current solution (before this code change) allows parameters to be
overwritten during a model run.  This causes subsequent model runs to
be initialized with the wrong parameters.  This happens for example when
one of the parameters is warehouse stock.  If the first model modifies the
amount of stock in the warehouse, then the second model run would be
initialized with that modified value.