Use click and add `mesa run`
This runs the project in the same way a user typically would, by
changing to the project directory (so that resources can be located
using relative paths).  However, rather than executing the `run.py`
directly, it assumes that importing `run` will produce a module with a
`server` attribute.  For the moment, it calls `server.launch()` but it
could be modified to call other methods or pass parameters to that
server, such as to run "headless"