fix issue 612 (overlapping viz elements when changing size) (#615)
an issue with overlapping vizualization elements was caused due to the
way an interaction handler was implimented. In that implimentation, two
canvas elements were overlayed on top of eachother using absolute
positioning with subsequent elements being positioned with hardcoded
values.

This removed all those hardcoded values and keeps the interaction
handler functioning by wrapping both the canvas elements that the
interaction handler uses in a single div. This also helps maintain
flexability by keeping the visualization element contained in one html
element.

PS. this change was applied to both the standard grid and the hexagonal
grid.