fix issue 765: D3 Network visualization overriding itself
The D3 network visualization was creating a whole new set of objects
each render and stacking it on top of the previous. This was wasteful,
unperformant, and poor D3.js style.

This change restructures that D3 code to take advantage of D3's update
model. Instead of creating new objects every render, the code now edits
existing objects, adding and removing as needed. (This is the intended
way to write D3.js code and the library makes this pattern very easy.)

This change also fixes a possible future issue where adding multiple svg
elements to a viz would break this one.