Add information on image as agent portrayal shape
At first I was wondering if it was possible to fill the grid with a custom image, but this was not clearly stated. Just recently I found out that it was already implemented! It is a shame not to use this feature. Could you consider also adding it to CanvasGridVisualization.py?