Docs: Add back some comments in space.py
Improve the space.py docs by adding back some comments.

Apparently, starting directly with "Args:" results in the function not being included in the docs. This fixes it by always starting with a short description (at least in space.py).

Also make the whitelines at the begin and end of comments consistent.