Adding logic to check whether there is agent data
There was a bug where the model output was only produced
if there were agent_reporters.

This is because the output data was created in a list
comprehension that ran once for every agent in
agent_reporters.

The new logic checks if there are agent reporters;
if not, it just outputs the model_reporters.