Move metric aggregation function to datacollector
Moves the metric aggegration code to the datacollector, in a new function get_agent_metric(). It returns a single value from the agent variables, using either a function in the statistics module or one of the built-in functions "min", "max", "sum" or "len".

In the ChartModule an optional "Metric" can now be called, which defaults to "mean".