Configure Black with a max line length of 100 chars
Adds a pyproject.toml file that configures Black with a maximum line length of 100 characters, up from the default of 88.

In many cases, this makes code more readable by allowing assignments and expressions to stay on a single line. It also encourages the use of longer, more descriptive variable names, which can make code easier to understand.