release CI: Run on all pushes and PRs, only publish on tag
Currently the release CI only run when a GitHub Release is created. This PR modifies that is runs on each PR and push, and uploads.

It uses the official action from PyPI: https://github.com/pypa/gh-action-pypi-publish

It also now uses build to build the wheel, instead of calling setup.py directly, which is deprecated.