ci: Speed up pip install by caching deps install output
Based on the benchmark, it speeds up by about 2x for the Ubuntu instances. While extracting the cache tarball for the Windows instance could be further sped up, with a `tar` binary that supports `zstd`.
See https://github.com/rht/mesa/pull/12.

The cache should be automatically cleared every month. Or alternatively, create a lockfile (either via Poetry of `pip-tools`)  that is automatically updated via Dependabot. Regardless, this is out of the scope of this PR.