benchmarks.yml: Another attempt at fixing PR branch checkout
This commit addresses the issue where GitHub Actions incorrectly checked out the main branch of the base repository when triggered by `issue_comment` events, instead of the correct branch from a forked repository. The solution involves using a GitHub script to fetch the pull request details associated with the issue comment, extracting the branch and repository information of the head (fork). This data is then used to explicitly checkout the correct branch from the forked repository. This ensures that the workflow operates on the intended branch, regardless of whether it's triggered by `pull_request_target` or `issue_comment` events, thereby maintaining the correct workflow context and enhancing security.