Ensure grasspatches info is only collected if gras is true (#181)
the datacollector allways includes grasspatches even if grass is False. This is breaking (the proposed changes to agent storage in Model.

The underlying problem is that Model.get_agents_of_type now raises a KeyError if type does not exist. I think this is desirable behavior becuase errrors should not be passed over in silence. But this requires fixing the wolf sheep example as done here.