Include examples in readthedocs (port) (#2392)
This updates conf.py to create markdown files for including the examples in the documentation. It also cleans up some warning on RTD, which happened while debugging the code. In this process, this also fixes #2360.

It currently only covers the basic examples because the advanced examples are not fully cleaned yet. 

---------

Co-authored-by: Jan Kwakkel <j.h.kwakkel@tudelft.nl>