Stabilize PropertyLayer
Removes the warning that the PropertyLayer is experimental, and thus stabilize it. It has been used in example models, integrated into the visualisation (both standalone and as part of the cell space) and in Mesa for almost a year without needing API changes.

Note that while the PropertyLayer itself is stable, not all it's integration might be, like the integration into the still experimental cell space.