Update boid_flockers example include flight direction (#2696)
I updated boid_flockers so that the direction of movement of agents is shown in the simulation.

I did this by changing the marker to caret-down (a triangle with a larger base so that the direction of motion is not ambiguous) and then rotating the marker according to the direction of motion of the agent.